routes: move turnilo routes to dedicated router
This commit moves turnilo dashboard routes to a dedicated module.
